+++ main/src/main/java/net/oneandone/stool/cli/Cleanup.java
@@ -15,13 +15,12 @@
  */
 package net.oneandone.stool.cli;
 
-import net.oneandone.stool.locking.Mode;
 import net.oneandone.stool.stage.Reference;
 import net.oneandone.stool.util.Server;
 
 public class Cleanup extends StageCommand {
     public Cleanup(Server server) {
-        super(server, Mode.NONE, Mode.EXCLUSIVE);
+        super(server);
     }
 
     @Override