+++ main/src/main/java/net/oneandone/stool/cli/Detach.java
@@ -16,7 +16,6 @@
 package net.oneandone.stool.cli;
 
 import net.oneandone.inline.ArgumentException;
-import net.oneandone.stool.locking.Mode;
 import net.oneandone.stool.util.Project;
 import net.oneandone.stool.util.Server;
 import net.oneandone.sushi.fs.file.FileNode;
@@ -25,7 +24,7 @@ import java.io.IOException;
 
 public class Detach extends ProjectCommand {
     public Detach(Server server, FileNode project) {
-        super(server, Mode.EXCLUSIVE, project);
+        super(server, project);
     }
 
     @Override