+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Stages are now the central part; multiple project can be associated with the sam
   * dumped svn credentials handling
   * dumped macros
 
+* removed client-side locking
+
 * added `running` field
 * changed server.xml: set `deployXML` to false 
   (since we now have exactly one web application per Tomcat, crossContext has no meaning;